extract project creation test helper
